[WFCORE-2564] Customize resolve-path to handle absolute paths
add checkAbsolutePath option to the resolve-path builder so that a
resource can customize the operation to ignore the relative-to value if
the path value corresponds to an absolute path.

The default builder does not check the absolute path (no changes to
current behaviour by default).

JIRA: https://issues.jboss.org/browse/WFCORE-2564